% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDashes.R
\name{addDashes}
\alias{addDashes}
\title{addDashes (for formatting code)}
\usage{
addDashes()
}
\value{

}
\description{
Adds '### --' (50 '-') for making sections in code and general formatting. The function is called by Rstudio up on a hot key combinations.
}
