% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDashes.R
\name{addHashes}
\alias{addHashes}
\title{addHashes (for formatting code)}
\usage{
addHashes()
}
\value{

}
\description{
Adds '###' (54 '#') for making sections in code and general formatting. The function is called by Rstudio up on a hot key combinations.
}
