% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_dynamic_cutoff.R
\name{apply_dynamic_cutoff}
\alias{apply_dynamic_cutoff}
\title{applies dynamic cutoff for selecting IR based on abundance of
  reference introns that varies according to sequencing depth.}
\usage{
apply_dynamic_cutoff(
  mat,
  refIntrons,
  dynamicCutoff = 1000,
  rowLogic = "and",
  verbose = F
)
}
\arguments{
\item{mat}{any value matrix, commonly splice exact or (splice left + splice right).}

\item{refIntrons}{reference introns/stably spliced exons identified from \emph{identify_ref_introns()}.}

\item{dynamicCutoff}{How abundant should IR be compared to refIntrons? default is 1000-times less abundant than refIntrons in each sample.}

\item{rowLogic}{logical operator to collapse 2D matrix into 1D. default is \emph{and}.}

\item{verbose}{T or F}
}
\value{
1D vector of IR meeting dynamic cutoff.
}
\description{
Sequencing depth for sample could vary tremendously; therefore,
  some IR may not pass fixed filter such as  splice exact must be >50 in all
  samples purely due to low seq depth for some samples. To solve this,
  selection criteria to check abundance of IR (by splice exact or SL + SR)
  should be varied based on sequencing depth. This function attempts to
  correct bias in sequencing depth when selecting IR events by looking for
  stably spliced exons that act as reference genes as it is done in qPCR (see
  \emph{identify_ref_introns()}), and then vary IR selection criteria based
  on seq counts of stable reference exons. Please note that
  \emph{identify_ref_introns()} should be run prior to running this function
  to identify reference spliced exons.
}
