% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_GenomicRange.R
\name{as.GenomicRange}
\alias{as.GenomicRange}
\title{converts 'chrX:23232-23232' or IRfinder output rowname to GRange}
\usage{
as.GenomicRange(coord)
}
\arguments{
\item{coord}{coord to be converted to genomic ranges}
}
\value{

}
\description{
converts 'chrX:23232-23232' or IRfinder output rowname to GRange
}
\examples{
as.GenomicRange("chrX:23232-23240")
}
