% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCoordWidth.R
\name{calculateCoordWidth}
\alias{calculateCoordWidth}
\title{calculates width/length of genomic coordinates}
\usage{
calculateCoordWidth(coord, unit = "kb")
}
\arguments{
\item{coord}{genomic coord in format: 'chr11:65083723-65083932'}

\item{unit}{NA or kb}
}
\value{
intron widths
}
\description{
calculates width/length of genomic coordinates
}
