% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMinProp.R
\name{checkMinProp}
\alias{checkMinProp}
\title{evaluates whether at least X proportion of samples in raw value matrix
  meets some threshhold value.}
\usage{
checkMinProp(mat, value, min_proportion, byrow = T)
}
\arguments{
\item{mat}{value matrix}

\item{value}{threshhold value to check in *mat*}

\item{min_proportion}{threshhold minimum proportions of samples (colnames) to
higher than the value parameter}

\item{byrow}{whether to calculate min_proportion by row}
}
\value{

}
\description{
This function is useful when selecting whether 50%
  (\emph{min_proportion}) of samples (\emph{mat}) has value more than
  \emph{value}.
}
