% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_gene_exp_IR_events.R
\name{corr_gene_exp_IR_events}
\alias{corr_gene_exp_IR_events}
\title{Correlate gene expression with number of IR events in samples}
\usage{
corr_gene_exp_IR_events(mat_gene_exp, mat_IR)
}
\arguments{
\item{mat_gene_exp}{gene expression matrix in genes x sample format.}

\item{mat_IR}{1D matrix of number of IR events in sample x num_IR_events format.}
}
\value{

}
\description{
Correlate gene expression with number of IR events in samples
}
