% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeqDataSetFromPreCalculatedMatrix.R
\name{DESeqDataSetFromPreCalculatedMatrix}
\alias{DESeqDataSetFromPreCalculatedMatrix}
\title{Generates valide DESeq2 Object to identify differentially retained
  introns}
\usage{
DESeqDataSetFromPreCalculatedMatrix(
  expDesign,
  designFormula,
  selectedIR = NA,
  mat_intronDepth = NULL,
  mat_spliceExact = NULL
)
}
\arguments{
\item{expDesign}{Valid experimental design to be passed to DESeq2}

\item{designFormula}{Valid design formula to be passed to DESeq2.}

\item{selectedIR}{a subset of IR as character vector for subsequent analysis.}

\item{mat_intronDepth}{matrix containing intron depth}

\item{mat_spliceExact}{matrix containing splice exact.}
}
\value{
DESeq2 object
}
\description{
This function generates DESeq2 object that will be passed to
  DESeq2 to identify differentially retained introns in biological conditions
  using GLM/linear models. Please note that raw intron depth and splice exact
  exported from IRFinder are used in calculating IR ratio using DESeq2. See \link{differential_IR} to calculate IR ratio.
}
