% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_IRF_output.R
\name{diagnose_IRF_output}
\alias{diagnose_IRF_output}
\title{returns values of key columns of the IRFinder output for selected samples and IRs.}
\usage{
diagnose_IRF_output(all, ir_index, selectSamples)
}
\arguments{
\item{all}{the mighty \emph{all} variable containing list of IRFinder columns.}

\item{ir_index}{index values of IRs to look}

\item{selectSamples}{samples}
}
\value{

}
\description{
returns values of key columns of the IRFinder output for selected samples and IRs.
}
