% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractGenomicCoord.R
\name{extractGenomicCoord}
\alias{extractGenomicCoord}
\title{Extracts genomic loci from irnames}
\usage{
extractGenomicCoord(irnames)
}
\arguments{
\item{irnames}{output of IRFinder in 'ESPN/ENSG00000187017/anti-near/1:  6425249-  6428225:+' format}
}
\value{
importFrom purrr map_chr
}
\description{
Extracts genomic loci from irnames: ESPN/ENSG00000187017/anti-near/1:  6425249-  6428225:+ to 'chr1:6425249-6428225'
}
