% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterSamples.R
\name{filterSamples}
\alias{filterSamples}
\title{Filters IR events based on flexible selection criteria.}
\usage{
filterSamples(criteria, sampleList, rowLogic = "and")
}
\arguments{
\item{criteria}{Text of of the criteria to be
applied on each sample in SampleList, termed '\emph{x}' in the criteria.}

\item{sampleList}{samples to apply the criteria to. A character vector.}

\item{rowLogic}{how to collapse 2D matrix into 1D: by \emph{and} or \emph{or}
operator. Alternatively, \emph{none} will return raw logical matrix with
same dimension as sampleList.}
}
\value{
1D or 2D logical vector representing samples passing the filtering criteria.
}
\description{
Flexible function that applies a selection criteria to each
  sample in the matrix for filtering IR. \emph{criteria} is directly
  evaluated by R; therefore, it must be a valid expression. \emph{criteria}
  \strong{must} contain a dummy variable \emph{x} that is used a placeholder for iteration.
}
\examples{
\dontrun{
criteria <- "all[[\"IR\"]][, x] > 0.1" # [,x] will be replaced by each item of sampleList.
sampleList <- colnames(all[["IR"]]) # for all samples in matrix containing IR values, all[["IR"]].

filtered_ir <- filterSamples(criteria, samples, "and" )

## above command will return IR events (rows in all variable) that have value of >0.1 in all samples provided.

}


}
