% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneID2Sym.R
\name{geneID2Sym}
\alias{geneID2Sym}
\title{Convers ensembl id to gene symbol.}
\usage{
geneID2Sym(ids)
}
\arguments{
\item{ids}{Ensembl ids to be converted to gene symbols}
}
\value{

}
\description{
Convers ensembl id to gene symbol.
}
