% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSym2ID.R
\name{geneSym2ID}
\alias{geneSym2ID}
\title{Convers gene symbol to ensembl id.}
\usage{
geneSym2ID(symbols)
}
\arguments{
\item{symbols}{gene symbols to be converted to Ensembl IDs.}
}
\value{

}
\description{
Convers gene symbol to ensembl id.
}
