% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_common_reference_genes.R
\name{get_common_reference_genes}
\alias{get_common_reference_genes}
\title{returns a list of common reference genes used in identifying stably spliced exons.}
\usage{
get_common_reference_genes()
}
\value{

}
\description{
returns a list of common reference genes used in identifying stably spliced exons.
}
