% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_IR_host_genes.R
\name{get_IR_host_genes}
\alias{get_IR_host_genes}
\title{Extracts gene symbols or id from IRFinder output names}
\usage{
get_IR_host_genes(ir, returnGeneSymbols = F)
}
\arguments{
\item{ir}{A vector of IR names exported from IRFinder in the format: symbol/EnsemblID/intron coordinates/strand.}

\item{returnGeneSymbols}{whether to return gene symbols (names, field 1) or gene id (field 2).}
}
\value{

}
\description{
Extracts gene symbols or id from IRFinder output names
}
