% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_ref_introns.R
\name{identify_ref_introns}
\alias{identify_ref_introns}
\title{identifies most stably spliced exons (reference for introns) in reference
  genes.}
\usage{
identify_ref_introns(se, ref_genes, multiRef = NA)
}
\arguments{
\item{se}{value matrix. could be splice exact or (splice left + splice right).}

\item{ref_genes}{character vector of reference genes}

\item{multiRef}{whether to consider multiple reference genes. User should first run the function without multiRef, select exons that are stable and then supply them as a list.}
}
\value{
list of stable intron names and their average counts in the data set
}
\description{
identifies most stably spliced exons (reference for introns) in reference
  genes. This information is then used to calculate theshhold in splice exact
  etc for abundance, ie retained introns should be expressed at 1000th of
  whatever expression of GAPDH (most stable reference intron) is. The algorithm used in identifying stable exons is similar to geNorm. Most
  genes have multiple exons, so only exons with highest eplice exacts are
  selected. The method calculates distance of the most abundant exon for each
  reference gene from the geomatric mean of all reference exons and then
  looks at which exon shows least variation from geomatric mean in all
  samples.
}
