% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intron_length.R
\name{intron_length}
\alias{intron_length}
\title{Returns length of introns in nucleotides}
\usage{
intron_length(irnames)
}
\arguments{
\item{irnames}{rownames of IRFinder output, eg 'HES3/ENSG00000173673/anti-near/1:6244241-6244350:+'.}
}
\value{

}
\description{
Returns length of introns in nucleotides
}
