% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intron_type.R
\name{intron_type}
\alias{intron_type}
\title{Returns type of intron retention event (aka clean or anti-near etc; 3rd field of IRFinder rownames)}
\usage{
intron_type(irnames)
}
\arguments{
\item{irnames}{rownames of IRFinder output, eg 'HES3/ENSG00000173673/anti-near/1:6244241-6244350:+'.}
}
\value{

}
\description{
Returns type of intron retention event (aka clean or anti-near etc; 3rd field of IRFinder rownames)
}
