% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isBalancedSpliced.R
\name{isBalancedSpliced}
\alias{isBalancedSpliced}
\title{Returns a logical vector checking whether a retained intron is balanced splicing}
\usage{
isBalancedSpliced(SL, SR, lowest = 0.8, highest = 1)
}
\arguments{
\item{SL}{splice left matrix}

\item{SR}{splice right matrix}

\item{lowest}{lowest SL:SR ratio}

\item{highest}{highest SL:SR ratio}
}
\value{
logical vector
}
\description{
Some retained introns have higher splice exact values but this single value does not tell whether it splicing or exon skipping. In normal splicing, splice left, splice right and splice exact are roughtly similar. However, exon skipping events results in uneven splice left and splice right reads, resulting their ratio swayed away from 1 that can be used to determine normal splicing versus exon skipping.
}
