% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPackages.R
\name{loadPackages}
\alias{loadPackages}
\title{Loads multiple packages if already installed}
\usage{
loadPackages(packages)
}
\arguments{
\item{packages}{packages to load as character vector.}
}
\value{
Null or errors showing missing packages
}
\description{
Loads multiple packages if already installed
}
