% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRawIR.R
\name{plotRawIR}
\alias{plotRawIR}
\title{plots raw IR, spliceLeft etc for a given IR.}
\usage{
plotRawIR(select_ir, ...)
}
\arguments{
\item{select_ir}{selecte_ir to plot. must be more than one AND match rownames or row number of all.}

\item{...}{optional parameters that will be passed to boxplot().}
}
\value{

}
\description{
plots raw IR, spliceLeft etc for a given IR.
}
