% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_IRF_warnings.R
\name{recode_IRF_warnings}
\alias{recode_IRF_warnings}
\title{Recodes IRF warning as numeric values. Recoding does not make a difference in memory storage.}
\usage{
recode_IRF_warnings(mat)
}
\arguments{
\item{mat}{matrix containing IRF warnings}
}
\value{

}
\description{
Recodes IRF warning as numeric values. Recoding does not make a difference in memory storage.
}
