% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceZero.R
\name{replaceZero}
\alias{replaceZero}
\title{replaces 0s in a matrix with lowest value.}
\usage{
replaceZero(temp)
}
\arguments{
\item{temp}{input matrix}
}
\value{
matrix
}
\description{
replaces 0s in a matrix with lowest value.
}
