% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRF_WARNING_CODES.R
\name{recode_irf_warnings}
\alias{recode_irf_warnings}
\title{Recodes IRF warning as numeric values.}
\usage{
recode_irf_warnings(mat)
}
\arguments{
\item{mat}{matrix containing IRF warnings}
}
\value{

}
\description{
Recodes IRF warning as numeric values.
}
